rmat generator use only float (removes spurious conversions)

+++ generator.h
@@ -72,19 +72,19 @@ class Generator {
   }
 
   EdgeList MakeRMatEL() {
-    const float A = 0.57, B = 0.19, C = 0.19;
+    const float A = 0.57f, B = 0.19f, C = 0.19f;
     EdgeList el(num_edges_);
     #pragma omp parallel
     {
       std::mt19937 rng;
-      std::uniform_real_distribution<float> udist(0, 1.0);
+      std::uniform_real_distribution<float> udist(0, 1.0f);
       #pragma omp for
       for (int64_t block=0; block < num_edges_; block+=block_size) {
         rng.seed(kRandSeed + block/block_size);
         for (int64_t e=block; e < std::min(block+block_size, num_edges_); e++) {
           NodeID_ src = 0, dst = 0;
           for (int depth=0; depth < scale_; depth++) {
-            double rand_point = udist(rng);
+            float rand_point = udist(rng);
             src = src << 1;
             dst = dst << 1;
             if (rand_point < A+B) {
